Change package name of util and api classes
Also add doc comments
