+++ app/src/main/java/com/starry/myne/api/models/BookSet.kt
@@ -14,14 +14,20 @@
  * limitations under the License.
  */
 
-package com.starry.myne.repo.models
+package com.starry.myne.api.models
+
 
 import androidx.annotation.Keep
+import com.google.gson.annotations.SerializedName
 
-/** Extra info from google books API */
 @Keep
-data class ExtraInfo(
-    val coverImage: String = "",
-    val pageCount: Int = 0,
-    val description: String = ""
+data class BookSet(
+    @SerializedName("count")
+    val count: Int,
+    @SerializedName("next")
+    val next: String?,
+    @SerializedName("previous")
+    val previous: String?,
+    @SerializedName("results")
+    val books: List<Book>
 )
