+++ app/src/main/java/com/starry/myne/api/models/Translator.kt
@@ -14,20 +14,18 @@
  * limitations under the License.
  */
 
-package com.starry.myne.repo.models
+package com.starry.myne.api.models
 
 
 import androidx.annotation.Keep
 import com.google.gson.annotations.SerializedName
 
 @Keep
-data class BookSet(
-    @SerializedName("count")
-    val count: Int,
-    @SerializedName("next")
-    val next: String?,
-    @SerializedName("previous")
-    val previous: String?,
-    @SerializedName("results")
-    val books: List<Book>
+data class Translator(
+    @SerializedName("name")
+    val name: String = "N/A",
+    @SerializedName("birth_year")
+    val birthYear: Int,
+    @SerializedName("death_year")
+    val deathYear: Int
 )
