+++ app/src/main/java/com/starry/myne/database/MyneDatabase.kt
@@ -26,7 +26,7 @@ import com.starry.myne.database.library.LibraryDao
 import com.starry.myne.database.library.LibraryItem
 import com.starry.myne.database.reader.ReaderDao
 import com.starry.myne.database.reader.ReaderData
-import com.starry.myne.utils.Constants
+import com.starry.myne.helpers.Constants
 
 @Database(
     entities = [LibraryItem::class, ReaderData::class],
