+++ app/src/main/java/com/starry/myne/helpers/Constants.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.starry.myne.utils
+package com.starry.myne.helpers
 
 object Constants {
     // Database
