+++ app/src/main/java/com/starry/myne/helpers/NetworkObserver.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.starry.myne.utils
+package com.starry.myne.helpers
 
 import android.content.Context
 import android.net.ConnectivityManager
