+++ app/src/main/java/com/starry/myne/helpers/book/BookLanguage.kt
@@ -1,4 +1,4 @@
-package com.starry.myne.utils.book
+package com.starry.myne.helpers.book
 
 import androidx.annotation.Keep
