+++ app/src/main/java/com/starry/myne/ui/screens/categories/composables/CategoryDetailScreen.kt
@@ -55,6 +55,9 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import coil.annotation.ExperimentalCoilApi
 import com.starry.myne.R
+import com.starry.myne.helpers.NetworkObserver
+import com.starry.myne.helpers.book.BookLanguage
+import com.starry.myne.helpers.book.BookUtils
 import com.starry.myne.ui.common.BookItemCard
 import com.starry.myne.ui.common.BookLanguageButton
 import com.starry.myne.ui.common.CustomTopAppBar
@@ -64,9 +67,6 @@ import com.starry.myne.ui.common.ProgressDots
 import com.starry.myne.ui.navigation.Screens
 import com.starry.myne.ui.screens.categories.viewmodels.CategoryViewModel
 import com.starry.myne.ui.theme.pacificoFont
-import com.starry.myne.utils.NetworkObserver
-import com.starry.myne.utils.book.BookLanguage
-import com.starry.myne.utils.book.BookUtils
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import java.util.Locale
