+++ app/src/main/java/com/starry/myne/ui/screens/detail/composables/BookDetailScreen.kt
@@ -83,15 +83,15 @@ import com.airbnb.lottie.compose.animateLottieCompositionAsState
 import com.airbnb.lottie.compose.rememberLottieComposition
 import com.starry.myne.MainActivity
 import com.starry.myne.R
+import com.starry.myne.helpers.Utils
+import com.starry.myne.helpers.book.BookUtils
+import com.starry.myne.helpers.getActivity
 import com.starry.myne.ui.common.BookDetailTopUI
 import com.starry.myne.ui.common.NetworkError
 import com.starry.myne.ui.common.ProgressDots
 import com.starry.myne.ui.screens.detail.viewmodels.BookDetailViewModel
 import com.starry.myne.ui.theme.figeronaFont
 import com.starry.myne.ui.theme.pacificoFont
-import com.starry.myne.utils.Utils
-import com.starry.myne.utils.book.BookUtils
-import com.starry.myne.utils.getActivity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
