+++ app/src/main/java/com/starry/myne/ui/screens/home/composables/HomeScreen.kt
@@ -85,6 +85,9 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import coil.annotation.ExperimentalCoilApi
 import com.starry.myne.R
+import com.starry.myne.helpers.NetworkObserver
+import com.starry.myne.helpers.book.BookLanguage
+import com.starry.myne.helpers.book.BookUtils
 import com.starry.myne.ui.common.BookItemCard
 import com.starry.myne.ui.common.BookLanguageButton
 import com.starry.myne.ui.common.NetworkError
@@ -94,9 +97,6 @@ import com.starry.myne.ui.screens.home.viewmodels.HomeViewModel
 import com.starry.myne.ui.screens.home.viewmodels.UserAction
 import com.starry.myne.ui.theme.figeronaFont
 import com.starry.myne.ui.theme.pacificoFont
-import com.starry.myne.utils.NetworkObserver
-import com.starry.myne.utils.book.BookLanguage
-import com.starry.myne.utils.book.BookUtils
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
