+++ app/src/main/java/com/starry/myne/ui/screens/library/composables/LibraryScreen.kt
@@ -103,15 +103,15 @@ import com.starry.myne.BuildConfig
 import com.starry.myne.MainActivity
 import com.starry.myne.R
 import com.starry.myne.database.library.LibraryItem
+import com.starry.myne.helpers.Utils
+import com.starry.myne.helpers.getActivity
+import com.starry.myne.helpers.isScrollingUp
 import com.starry.myne.ui.common.CustomTopAppBar
 import com.starry.myne.ui.navigation.Screens
 import com.starry.myne.ui.screens.library.viewmodels.ImportStatus
 import com.starry.myne.ui.screens.library.viewmodels.LibraryViewModel
 import com.starry.myne.ui.screens.settings.viewmodels.ThemeMode
 import com.starry.myne.ui.theme.figeronaFont
-import com.starry.myne.utils.Utils
-import com.starry.myne.utils.getActivity
-import com.starry.myne.utils.isScrollingUp
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import me.saket.swipe.SwipeAction
