+++ app/src/main/java/com/starry/myne/ui/screens/welcome/viewmodels/WelcomeViewModel.kt
@@ -19,7 +19,7 @@ package com.starry.myne.ui.screens.welcome.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.starry.myne.utils.PreferenceUtil
+import com.starry.myne.helpers.PreferenceUtil
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
